+++ api/application/routes/login.js
@@ -1,0 +1,17 @@
+const { initiateLogin, completeLogin } = require('../controllers/login');
+
+exports.register = async function(server) {
+  server.route([
+    {
+      method: "POST",
+      path: "/login",
+      handler: initiateLogin
+    },
+    {
+      method: "GET",
+      path: "/complete-login",
+      handler: completeLogin,
+    }
+    // test user data link for authentication?
+  ]);
+};
