+++ api/application/routes/sign-up.js
@@ -1,0 +1,9 @@
+const { signUp } = require("../controllers/sign-up");
+
+exports.register = async function(server) {
+  server.route([{
+    method: "POST",
+    path: "/sign-up",
+    handler: signUp
+  }]);
+};
