+++ api/application/routes/subscribe-beta-tester.js
@@ -1,0 +1,9 @@
+const { subscribeBetaTester } = require('../controllers/subscribe-beta-tester');
+
+exports.register = async function(server) {
+  server.route([{
+    method: 'POST',
+    path: '/subscribe-beta-tester',
+    handler: subscribeBetaTester
+  }]);
+}
